don't spam user with warning messages.
this commit is related to #1898 and make sure failures in gene cluster
alignment step is reported only once and not for each gene cluster. the
`--debug` flag places raw sequences somewhere for the user to be able to
go back to them to see what may have gone wrong.

plus, we now pass around the gene cluster names that failed alignmet,
thus, the homogeneity index calculations do not have to print out their
own warning for something the user is already warned.

overall cleaner output with no impact on function.