a heuristic to fix the snakemake error for missing `--cores`
Here is what snakemake had to say for `anvi-run-workflow` runs that did
not include `--cores XX` through the `--additional-params` mechanism:

Error: you need to specify the maximum number of
CPU cores to be used at the same time. If you want
to use N cores, say --cores N or -c N. For all cores
on your system (be sure that this is appropriate)
use --cores all. For no parallelization use --cores 1
or -c 1.