sanity check for parsed `args`.
This was necessary because of a cryptic error reported by Hilary
Morrison. She set the number of threads to -40 rather than 40, and got a
very weird error way later in her workflow. This should fix such typos
from the get go and we can expand with additional simple sanity checks