consider all possible merges
this is a much more comprehensive version of the previous way of merging
stretches that ensures merging is finalized when all possible merges are
tested. This change will influence @FlorianTrigodet's inversions, but I
do not expect to see a significant change. Although, it would be great
to test it to ensure results don't change noticably.