not only dbs, but all anvi'o artifacts
this program was written only from the perspective of anvi'o dbs. but it
also migrates JSON files, and perhaps ohters in the future. this issue
was most obvious with the command line flags, `--migrate-dbs-safely` and
`--migrate-dbs-quickly`. which are weird to pass when the user is
migrating JSON files. now the code and parameters speak of 'artifacts'
when necessary, and not dbs.