don't list 'hypothetical protein' as function
this is a rather important change. so far if the function is not known,
we are setting PGAP functio to hypothetical protein instead of None.
which is not the most elegant thing to do to protect downstream analyses
from reduntant code to remove those genes with unknown functions from
analyses of function.