a significant change
the code surrounds this line is from 2015. read it again: TWO THOUSAND AND FIFTEEN.

This change is necessary NOW, because the view data for functions display can be incomplete (i.e., the views table will not have an entry if a given function is not in a given genome).

this results in 'key not found' error here, obviously. this has never been a problem, because we stored entries for every split in every metagenome before this -- so this line worked all this time. actually this could be a GREAT way to reduce size. if a split doesn't have any coverage in a metageome, it doesn't have to be in the view table. I'm not sure how I didn't think of this before.

BUT IT IS A LOT OF WORK NOW, SO CONTINUE TO SAY GOODBYE TO YOUR STORAGES I GUESS.