an entirely different approach for usability
starting with FASTA files have advantages, but I realized that starting
with contigs-db files will be even better. for instance, if you want to
run this tool with many genomes with different thresholds, there will be
as many times of `anvi-run-kegg-kofams` when one should be sufficient :)

plus, there is no need to ask the user to provide a list of modules, or
the location for the KEGG data if they are OK fine with the defaults.
after these changes the script can be run on a contigs-db file without
any additional parameters.

In this revision I simply translated the code as much as I can to one
that actually can work with contigs-db files, and default values.