some critical changes to sanity check gene calls
this probably will be irrelevant to 99.9% of anvi'o projets, but in our
recent discussion with Zac, I realized the R. pom transcriptomic
profiles were having hard time to be summarized at the gene level. after
hours of frustruation, I realized that in fact some of the gene calls
that exists in the contigs-db, were not actually occurring in contigs,
therefore it was impossible for anvi'o to recover the coverages of these
genes. for the R. pom people, thse are gene calls 1598 and 1599.

With this change, anvi'o is prepared to handle even such extreme cases.