critical updates to get things to run
we now go with `protein_structure_informed_gene_clusters_dict`
tentatively as per the discussions in #lab.

once we have everything ready, meren will go through all the variable
names to try to fix the ones that are not quite working, and we will
open this up to the community input.