there is another definition of the same function
and it is more advanced, and less recent. so I have no clue how did this
happen. could it be really the case that the same function was
independently implemented in another branch and then thigns got merged?
really very spooky. if you like a detective story, git co to this
commit and git blame :p