associate people with their commits
there is an annoying issue of not being able to link people given their
GitHub usernames to their commits to the git repository of anvi'o. I
decided to keep track of this so all commits from a single individual
can be consolidated and every commit can be accounted for. I will have
some updates to the web page later :)