a new decorator to wrap every lazy loading func
This became a necessity since lazy loading simply ruined our progress calls for the terminal output, as the code in ContigsSuperclass no longer follows a linear path. For instance a progress instance starts, then the code reaches to a variable that needs to be load lazily, and in the function it loads it there is a `self.progress.new` call that really complicates everything. With this decorator + wrapper solution we are reaching in new depths in anvi'o codebase with technically sound hacks, and implement a logic that can work with every case of 'progress'. Well, fingers crossed.