function annotation tables are much better now
this is a lot of change in one go, but essentially these changes remove
a lot of redundancy by collapsing functions that show functions and
metabolic involvement for genes and gene clusters into a single one so
it is much easier to maintain them.

Plus, we now can actually choose what to display, and only keep function
annotation sources of interest in the final table, or remove entries
with no annotatation.