Merge JS improvements from pangeome_graph branch.'
A lot has happened to these files in multiple commits to improve
the JS code for bins and utils. Cherry-picking commits didn't work
because we don't make atomic commits and every commit is a 10,000
lines of code change :( so my only option was to do git show branch:file > file
which is really unprofessional, but whatever. it works.