new --dry-run parameter and function to print help
The source files and their expected locations on local computers is a menace to sort out. the new --dry-run parameter will help the users to figure those things out.

This change will address issues like #2480 by @MjelleLab and enable users to be able to download files and move them to the correct locations in offline computers before running `anvi-setup-ncbi-cogs`.

It also comes with some updates to the program docs.