fix collapsed node information retrieval
What was wrong? In the anvi'o interactive interface the user can
COMMAND-RIGHT-CLICK on any node in the tree shown in the center,
and click "collapse these items". In which case, the interface
would collapse those items into a triangular shape, and allow
user to save the new tree, upon which the collapse data would
also be saved with the tree in the additional column of the
item orders table in the profile database.

The problem was the following: when the user changed from one
tree to another in the interface, all collapse data would be
lost and would never come back.

It turns out this is due to a double-encoding of the JSON
formatted additional information. This commit fixes it
once and for all.