find the optimal ref start, not the first one.
 This is a better Optimal Start Position algorithm. Previously when the flag `--use-dnaa` not used, the code was trying to find *a* conserved region across all (or most) genomes to orient the selected reference to a biologically meaningful position. Bt the problem with that is that the first region found like that was not the most meaningful region.

With this commit I'm changing that behavior. The new algorithm first finds all conserved regions with no early termination, identifies all contiguous regions with over 80% genome coverage, selects the longest region, then calls genes on the reference, finds all intergenic regions within the longest conserved region, and finally selects the intergenic position closest to the middle of the conserved region so we don't split a gene :)

Then we rotate the reference to this location to start the reorientation of the remaining genomes.